Test whether logging works on ASCII locale
This also adds the en_US.UTF-8 locale to the testing Dockerfile, as the
locale of the user running the tests affects the outcome. The
user's locale is backed up before the tests are run, then reset
afterwards. While it is possible to the run other tests with
a non-existent locale, it is not possible to set a non-existent locale.

If we don't care what a potentially wrong locale does to other tests, we
could set the locale to "C.UTF-8" after the tests and ignore the test
runner's locale.